create AutoFitContentPlaceholder component
