+++ frontend/src/components/comments/CommentForm.jsx
@@ -1,12 +1,12 @@
 // NPM Packages
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState } from 'react';
+import AutoFitContentPlaceholder from '../AutoFitContentPlaceholder';
 
 export default function CommentForm({ post, onSubmit }) {
   // Local State
-  const [commentContentText, setCommentContentText] = React.useState('');
+  const [commentContentText, setCommentContentText] = useState('');
   const commentObject = require('../../assets/images/comment.svg');
   const commentURL = commentObject;
-  const textArea = useRef(null);
 
   // Methods
   const handleSubmit = () => {
@@ -17,33 +17,13 @@ export default function CommentForm({ post, onSubmit }) {
     setCommentContentText('');
   };
 
-  // Make textArea placeholder's height adaptive to the content
-  useEffect(() => {
-    const originalHeight = '52px';
-    textArea.current.style.height = originalHeight;
-    // counter 4px of padding included in scrollHeight
-    const newHeight = textArea.current.scrollHeight - 4;
-    textArea.current.style.height = `${newHeight}px`;
-    if (newHeight > 200) {
-      textArea.current.style.overflowY = 'scroll';
-    } else {
-      textArea.current.style.overflowY = 'none';
-    }
-  }, [commentContentText]);
-
-  const onTextAreaValueChange = (value) => {
-    setCommentContentText(value);
-  };
-
   return (
     <form>
       <div className="form-field">
-        <textarea
-          ref={textArea}
-          className="form-input"
+        <AutoFitContentPlaceholder
           placeholder="Write your comment here."
-          value={commentContentText}
-          onChange={(e) => onTextAreaValueChange(e.target.value)}
+          contentText={commentContentText}
+          setContentText={setCommentContentText}
         />
       </div>
       <div>
