+++ frontend/src/components/posts/PostForm.jsx
@@ -1,13 +1,12 @@
 // NPM Packages
-
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState } from 'react';
+import AutoFitContentPlaceholder from '../AutoFitContentPlaceholder';
 
 export default function PostForm({ onSubmit }) {
   // Local State
   const [contentText, setContentText] = useState('');
   const postObject = require('../../assets/images/post.svg');
   const postURL = postObject;
-  const textArea = useRef(null);
 
   // Methods
   const handleSubmit = () => {
@@ -17,24 +16,6 @@ export default function PostForm({ onSubmit }) {
     // Clear the input field
     setContentText('');
   };
-  
-  // Make textArea placeholder's height adaptive to the content
-  useEffect(() => {
-    const originalHeight = '52px';
-    textArea.current.style.height = originalHeight;
-    // counter 4px of padding included in scrollHeight
-    const newHeight = textArea.current.scrollHeight - 4;
-    textArea.current.style.height = `${newHeight}px`;
-    if (newHeight > 200) {
-      textArea.current.style.overflowY = 'scroll';
-    } else {
-      textArea.current.style.overflowY = 'none';
-    }
-  }, [contentText]);
-
-  const onTextAreaValueChange = (value) => {
-    setContentText(value);
-  };
 
   return (
     <div className="form-container">
@@ -42,12 +23,10 @@ export default function PostForm({ onSubmit }) {
         <div>
           <div>
             <div className="form-field">
-              <textarea
-                ref={textArea}
-                className="form-input"
+              <AutoFitContentPlaceholder
                 placeholder="What's on your mind?"
-                value={contentText}
-                onChange={(e) => onTextAreaValueChange(e.target.value)}
+                contentText={contentText}
+                setContentText={setContentText}
               />
             </div>
